This half-done piece...
Added the /vshop add command to add items to a shop;
Changed buyprice and sellprice to be per stack instead of per stackitem;

Wrote a custom inv wrapper to be able to set items where i want to set
them in the InvPrep ( I don't give a fick about offer if I want things
to be at a certain point, sorry );
Wrote some whacky code for the inventory clicked event to figure out if
the event happened inside the chest or player inventory ( not for anyone
to use! ); the buy function can't count for shit and won't withdraw
money? is this on my side or economy?